# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "School Hostel Enterprise API",
    "description": "Enterprise-grade API for school academics and residential hostel operations. It includes RBAC, students, attendance, fee collection, hostel occupancy, mess, visitors, leave approvals, complaints, dashboards, and audit logging.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/fees/invoices": {
      "post": {
        "tags": [
          "Fees & Payments"
        ],
        "summary": "Create Invoice",
        "operationId": "create_invoice_api_v1_fees_invoices_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvoiceCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Fees & Payments"
        ],
        "summary": "List Invoices",
        "operationId": "list_invoices_api_v1_fees_invoices_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "student_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Student Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/InvoiceStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvoiceRead"
                  },
                  "title": "Response List Invoices Api V1 Fees Invoices Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fees/invoices/{invoice_id}/payments": {
      "post": {
        "tags": [
          "Fees & Payments"
        ],
        "summary": "Pay Invoice",
        "operationId": "pay_invoice_api_v1_fees_invoices__invoice_id__payments_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "invoice_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Invoice Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/fees/summary": {
      "get": {
        "tags": [
          "Fees & Payments"
        ],
        "summary": "Fee Summary",
        "operationId": "fee_summary_api_v1_fees_summary_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InvoiceCreate": {
        "properties": {
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "fee_type": {
            "type": "string",
            "title": "Fee Type"
          },
          "billing_period": {
            "type": "string",
            "title": "Billing Period"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Amount"
          },
          "due_date": {
            "type": "string",
            "format": "date",
            "title": "Due Date"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "fee_type",
          "billing_period",
          "amount",
          "due_date"
        ],
        "title": "InvoiceCreate"
      },
      "InvoiceRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "invoice_no": {
            "type": "string",
            "title": "Invoice No"
          },
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "fee_type": {
            "type": "string",
            "title": "Fee Type"
          },
          "billing_period": {
            "type": "string",
            "title": "Billing Period"
          },
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount"
          },
          "paid_amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Paid Amount"
          },
          "due_date": {
            "type": "string",
            "format": "date",
            "title": "Due Date"
          },
          "status": {
            "$ref": "#/components/schemas/InvoiceStatus"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "invoice_no",
          "student_id",
          "fee_type",
          "billing_period",
          "amount",
          "paid_amount",
          "due_date",
          "status",
          "notes"
        ],
        "title": "InvoiceRead"
      },
      "InvoiceStatus": {
        "type": "string",
        "enum": [
          "pending",
          "partial",
          "paid",
          "overdue",
          "cancelled"
        ],
        "title": "InvoiceStatus"
      },
...
      "PaymentCreate": {
        "properties": {
          "amount": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Amount"
          },
          "payment_method": {
            "type": "string",
            "title": "Payment Method"
          },
          "transaction_ref": {
            "type": "string",
            "maxLength": 100,
            "minLength": 3,
            "title": "Transaction Ref"
          }
        },
        "type": "object",
        "required": [
          "amount",
          "payment_method",
          "transaction_ref"
        ],
        "title": "PaymentCreate"
      },
      "PaymentRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "invoice_id": {
            "type": "integer",
            "title": "Invoice Id"
          },
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount"
          },
          "payment_method": {
            "type": "string",
            "title": "Payment Method"
          },
          "transaction_ref": {
            "type": "string",
            "title": "Transaction Ref"
          },
          "received_by": {
            "type": "integer",
            "title": "Received By"
          },
          "paid_at": {
            "type": "string",
            "format": "date-time",
            "title": "Paid At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "invoice_id",
          "amount",
          "payment_method",
          "transaction_ref",
          "received_by",
          "paid_at"
        ],
        "title": "PaymentRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}